refactor: extract agent selection into internal/agentselect
Moves the agent-selection cluster out of the danger-zone runner.go into a
new, focused internal/agentselect package: Select (was selectAgents),
DropCLITwins, ProviderSpecsFromConfig, ExperimentalOnlyNames, ParseOnlyList
(exported because main.go / runMultiLLMReview call them), plus the
unexported isSunsetAndNotForced / applyConfig helpers.

Decoupling: selectAgents took *sharedFlags but only ever read sf.only, so
Select now takes a plain `only string` — removing the cmd-type dependency
and making the package import-clean (cli + config only, no cycle).

pickAgents stays in cmd (it owns the real CLI/provider detection I/O) and
now calls agentselect.{DropCLITwins,ProviderSpecsFromConfig,Select}; the
config-validate `--only` guard calls agentselect.ParseOnlyList; main.go's
experimental-only note calls agentselect.ExperimentalOnlyNames.

Behavior-preserving: all selection tests (default/only/sunset/config-
disable, dropCLITwins, parseOnlyList) moved with it and pass, rewritten to
pass `only` as a string instead of constructing a sharedFlags. runner.go
1430 → 1177 LOC (1544 → 1177 across the full Phase 2). New package at
73.1% coverage.
